pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["David-Roklem <roklem@mail.ru>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"
Flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.30.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
